Add extra check in decode. Split up steps so they're easier to manage. Add some initial auto-generated doco. Add doco to show example usage.
TODO: Add tests.
