+++ operators/vision/decode_image.cc
@@ -25,6 +25,10 @@ void KernelDecodeImage::Compute(OrtKernelContext* context) {
   const cv::Mat encoded_image(encoded_image_sizes, CV_8UC1, const_cast<void*>(encoded_image_data));
   const cv::Mat decoded_image = cv::imdecode(encoded_image, cv::IMREAD_COLOR);
 
+  if (decoded_image == nullptr) {
+    ORT_CXX_API_THROW("[DecodeImage] Invalid input. Failed to decode image.", ORT_INVALID_ARGUMENT);
+  };
+
   // Setup output & copy to destination
   const cv::Size decoded_image_size = decoded_image.size();
   const int64_t colors = 3;
